requirements.txt
torch
numpy
fastapi
uvicorn[standard]
google-cloud-storage
